src/backend/requirements.txt
# Core
fastapi>=0.104.1,<0.110.0
uvicorn[standard]>=0.24.0,<0.30.0
pydantic>=2.5.0,<2.10.0
pydantic-settings>=2.1.0,<2.5.0

# Database
sqlalchemy>=2.0.23,<2.0.36
asyncpg>=0.29.0,<0.30.0
alembic>=1.13.1,<1.14.0

# AI/ML
openai==1.5.0
langchain==0.1.0
langchain-community==0.0.13
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.4

# Git Integration
python-gitlab==4.1.1
PyYAML==6.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
structlog==23.2.0
prometheus-client==0.19.0
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0